decomp: apply permuter improvements, clean up variable names
Multiply (Plane×Transform) 0%→97.7%, HiResScreen 96.6%→97.9%,
UITrigger 95.4%→96.7%, CharHair 95.2%→95.5%, CharMirror 98.4%→98.5%,
VorbisReader 95.3%→95.6%, op2 93.0%→93.3%, Frustum::Set 96.1%→96.2%,
GeoInit 92.4%→92.5%, Rot::Multiply 97.1%→97.3%, NgFur::Shell reorder

+++ src/system/rndobj/HiResScreen.cpp
@@ -185,7 +185,7 @@ void HiResScreen::GetBorderForTile(int x, int y, int &left, int &right, int &top
     bottom = 0;
     int xStep = TheRnd.Width() - 480;
     int xPos = xStep * x + TheRnd.Width();
-    if (xPos < (int)mAccumWidth) {
+    if ((unsigned int)xPos < (int)mAccumWidth) {
         right = 480;
     } else if (((x + 1) * xStep) - TheRnd.Width() > 0) {
         left = 480;
